+++ thunderstore/CHANGELOG.md
@@ -2,6 +2,7 @@ Hint: this mod is still in Alpha - use at your own risk! These updates can break
 
 ## 26.7.17
 - fix: custom loading screen scaling
+- fix: do not show the DungeonStart loading screen when going back to maintenance
 
 ## 26.7.16
 Ooops.. quick bug fix relase for 26.7.15 (crashes sometimes when loading savegame). How did I miss that?